Fix weather crash when cloud_cover is returned as a float
Meteosource intermittently returns whole-number fields such as
cloud_cover as floats (e.g. 0.0), which failed to unmarshal into the int
field with 'cannot unmarshal number 0.0 into Go struct field
.current.cloud_cover of type int'. Make the response field a float64,
round it when building JSON output, and add a regression test covering a
float cloud_cover payload.

+++ morning-tasks/weather.go
@@ -13,7 +13,7 @@ type weatherResponse struct {
 	Current struct {
 		Summary     string  `json:"summary"`
 		Temperature float64 `json:"temperature"`
-		CloudCover  int     `json:"cloud_cover"`
+		CloudCover  float64 `json:"cloud_cover"`
 		Wind        struct {
 			Speed float64 `json:"speed"`
 			Dir   string  `json:"dir"`
@@ -80,7 +80,7 @@ func formatWeather(w io.Writer, data *weatherResponse, date time.Time) {
 	if isToday {
 		c := data.Current
 		_, _ = fmt.Fprintf(w, "Now: %s, %.0f°C\n", c.Summary, c.Temperature)
-		_, _ = fmt.Fprintf(w, "Wind: %.0f km/h %s | Clouds: %d%%\n", c.Wind.Speed, c.Wind.Dir, c.CloudCover)
+		_, _ = fmt.Fprintf(w, "Wind: %.0f km/h %s | Clouds: %.0f%%\n", c.Wind.Speed, c.Wind.Dir, c.CloudCover)
 		if c.Precipitation.Total > 0 {
 			_, _ = fmt.Fprintf(w, "Precipitation: %.1f mm (%s)\n", c.Precipitation.Total, c.Precipitation.Type)
 		}
